requirements.txt
streamlit>=1.32.0
openai>=1.12.0
python-dotenv>=1.0.0
chardet>=5.2.0
pandas>=2.0.0
plotly>=5.17.0
fpdf2>=2.7.6
Pillow>=10.0.0